🎨 refactor : ConfirmModal 컴포넌트로 댓글 삭제 기능 수정

+++ src/components/Comment.jsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect, useRef } from 'react';
 import axios from 'axios';
 import { Button, Card, Form, Dropdown, Modal } from 'react-bootstrap';
 import '../styles/Comment.css'
+import ConfirmModal from '../components/common/modals/ConfirmModal';
 
 const Comment = ({ comment, postId, user, created_at, onCommentChange }) => {
     const [isEdit, setIsEdit] = useState(false);
@@ -17,13 +18,19 @@ const Comment = ({ comment, postId, user, created_at, onCommentChange }) => {
         setShowModal(true);
     };
 
-    const handleConfirmDelete = async () => {
-        try {
-            await axios.delete(`http://localhost:3300/comments/${comment.id}`);
-            await onCommentChange();
+    const handleConfirmDelete = async (flag) => {
+        if (flag) {
+            // 사용자가 확인 버튼을 클릭했을 때 댓글 삭제 
+            try {
+                await axios.delete(`http://localhost:3300/comments/${comment.id}`);
+                await onCommentChange();
+                setShowModal(false);
+            } catch (error) {
+                console.error(error);
+            }
+        } else {
+            // 사용자가 취소 버튼을 클릭했을 때 모달 닫기
             setShowModal(false);
-        } catch (error) {
-            console.error(error);
         }
     };
 
@@ -36,7 +43,7 @@ const Comment = ({ comment, postId, user, created_at, onCommentChange }) => {
         }
 
         try {
-            await axios.put(`http://localhost:3300/comments/${comment.id}`, { 
+            await axios.put(`http://localhost:3300/comments/${comment.id}`, {
                 postId,
                 text: editedComment,
                 userId: user.userId,
@@ -57,7 +64,7 @@ const Comment = ({ comment, postId, user, created_at, onCommentChange }) => {
     const handleKeyDown = (e) => {
         if (e.key === 'Enter' && !e.shiftKey) { // 엔터키를 눌렀을 때 (Shift 키를 누르지 않았을 경우)
             e.preventDefault(); // 개행 방지
-            handleSave(e); 
+            handleSave(e);
         }
     };
 
@@ -79,7 +86,7 @@ const Comment = ({ comment, postId, user, created_at, onCommentChange }) => {
                             rows={editedComment.split('\n').length || 1}
                             ref={textareaRef}
                             value={editedComment}
-                            onChange={e => {setEditedComment(e.target.value);}}
+                            onChange={e => { setEditedComment(e.target.value); }}
                             onKeyDown={handleKeyDown} />
                         <Button variant='success' onClick={handleSave}>저장</Button>
                         <Button variant='secondary' onClick={handleCancel}>취소</Button>
@@ -88,27 +95,22 @@ const Comment = ({ comment, postId, user, created_at, onCommentChange }) => {
                     <div className='comment-container'>
                         <Card.Title>{comment.userId}</Card.Title>
                         <Card.Text>{comment.text}</Card.Text>
-                        <Card.Text>{comment.created_at  }</Card.Text>
+                        <Card.Text>{comment.created_at}</Card.Text>
                         {user.userId === comment.userId ? (
                             <Dropdown>
-                            <Dropdown.Toggle variant='primary' id='dropdown-basic'>더보기</Dropdown.Toggle>
-                            <Dropdown.Menu>
-                                <Dropdown.Item onClick={handleEdit}>수정</Dropdown.Item>
-                                <Dropdown.Item onClick={handleDelete}>삭제</Dropdown.Item>
-                            </Dropdown.Menu>
+                                <Dropdown.Toggle variant='primary' id='dropdown-basic'>더보기</Dropdown.Toggle>
+                                <Dropdown.Menu>
+                                    <Dropdown.Item onClick={handleEdit}>수정</Dropdown.Item>
+                                    <Dropdown.Item onClick={handleDelete}>삭제</Dropdown.Item>
+                                </Dropdown.Menu>
                             </Dropdown>
                         ) : ''}
-                        
-                        <Modal show={showModal} onHide={() => setShowModal(false)}>
-                            <Modal.Header closeButton>
-                                <Modal.Title>삭제 확인</Modal.Title>
-                            </Modal.Header>
-                            <Modal.Body>정말로 삭제하시겠습니까?</Modal.Body>
-                            <Modal.Footer>
-                                <Button variant='secondary' onClick={() => setShowModal(false)}>취소</Button>
-                                <Button variant='danger' onClick={handleConfirmDelete}>삭제</Button>
-                            </Modal.Footer>
-                        </Modal>
+                        <ConfirmModal
+                            onFlag={handleConfirmDelete}
+                            title='삭제 확인'
+                            text='정말 삭제하시겠습니까?'
+                            show={showModal}
+                        />
                     </div>
                 )}
             </Card.Body>
